💥 Move the FastAPI health router into grelmicro.integrations.fastapi
