+++ docs/changelog.md
@@ -6,6 +6,7 @@
 
 * 💥 Rename the recurring-task decorator from `@task.interval(seconds=...)` to `@task.every(seconds=...)`, pairing it with `@task.cron(...)` as a verb-form family. The `seconds=` keyword is unchanged. Update your call sites.
 * 💥 Move the framework integration modules into `grelmicro.integrations`. `grelmicro.fastapi` becomes `grelmicro.integrations.fastapi`, so `GrelmicroMiddleware` is now `from grelmicro.integrations.fastapi import GrelmicroMiddleware`, and the new FastStream wiring lives at `grelmicro.integrations.faststream`.
+* 💥 Move the FastAPI health router from `grelmicro.health.fastapi` to `grelmicro.integrations.fastapi`, so all FastAPI integration code (middleware, install, health router) lives under `grelmicro.integrations`. Update `from grelmicro.health.fastapi import health_router` to `from grelmicro.integrations.fastapi import health_router`.
 * 💥 Collapse the four `@task.every` lock passthroughs into one typed `lock=TaskLock(...)`. The `lease_duration`, `min_hold_duration`, `backend`, and `worker` parameters are removed. Pass a `TaskLock` instead, like `@task.every(seconds=60, lock=TaskLock(lease_duration=300))`. The lock keeps its default `"default"` name and is re-stamped to the task name, so you never repeat it. The lock's settings are authoritative. `leader=` and `sync=` stay separate, self-documenting parameters.
 * 💥 `HealthChecks` now namespaces its env vars per instance: the default instance keeps `GREL_HEALTH_*`, but a named instance reads `GREL_HEALTH_{NAME_UPPER}_*` (matching `Lock`, `CircuitBreaker`, and the other named components). Update env vars for any named `HealthChecks`.
 * 💥 The log filters now namespace env vars per instance. `GREL_DUPLICATE_FILTER_*` becomes `GREL_DUPLICATEFILTER_*` and `GREL_RATE_LIMIT_FILTER_*` becomes `GREL_RATELIMITFILTER_*`, with a new `env_name=` kwarg for a named instance (`GREL_DUPLICATEFILTER_{NAME}_*`). `DuplicateFilter.key_mode` also gains `logger`, `level`, and `global` so both filters share one vocabulary. Update env vars and `env_prefix=` overrides.
