+++ docs/reference/health.md
@@ -1,7 +1,7 @@
 # Health
 
 - **Start here**: [Health Checks guide](../health.md)
-- **FastAPI integration**: [`health_router`](../health.md) for liveness, readiness, and health endpoints.
+- **FastAPI integration**: [`health_router`](fastapi.md) for liveness, readiness, and health endpoints.
 
 ::: grelmicro.health
     options:
@@ -16,8 +16,3 @@
         - HealthReport
         - HealthStatus
         - get_health_checks
-
-::: grelmicro.health.fastapi
-    options:
-      members:
-        - health_router
